Changes

Documentation

data Change

Enum of the possible Changes to apply on the AST | Inspired by the list provided in the Seminal paper (2006, p. 4)

Constructors

GenericReplacement
Replace a variable with undefined

AddArg
Add arbitrary new argument to function call

DelArg
Delete argument to function call

SwapArg
Swap arguments to function call

ArrayToTuple
Turns an array into a tuple

TupleToArray
Turns a tuple into an array

ArgsToTuple
Turns application arguments into a tuple | Either all of them, or some

RemoveCaseMatch
Remove `case of` match

SetCaseToUndef
Set value of `case of` match to undefined

SetMatchToUndef
Set match of `case of` match to undefined

RemoveTypeSignature
Remove type signature

IfToTrue
Replace x in `if x then ... else ...` by True

LetToUndef
Set value of `let x = ` to undefined

WhereToUndef
Set one value in `where` clause to undefined

SingletonToItem
Turn `[x]` into x

ItemToSingleton
Turn x into `[x]`