Seminal

Description

Entrypoint to Seminal

Documentation

runSeminal :: FilePath -> IO Status

Run Seminal on a source file. If it returns Nothing, the file typechecks, otherwise, provides an ordered list of change suggestions

data Status

Constructors

Success
When the file typechecks without any changes

InvalidFile String
When an error occurs while parsing the file, unrelated to typechecking

Changes [Change]
An ordered list of change suggestions