Seminal.Change

Documentation

data Change node

Defines a change to apply on the AST. The namings are inspired by the astRepl (Seminal, 2006, p. 5)

newChange

Arguments

:: Outputable node

=> node
The node that is changed

-> node
The new node

-> SrcSpan
The location of the original change

-> [Change node]
Followups

-> Maybe String

-> ChangeType

-> Change node
Output

Exported constructor for Change. It will allow easier build of related doc

wrapLoc :: (SrcSpan -> a -> l) -> Change a -> Change l

Rewraps tha location to the change type

data ChangeDoc

Gives information about a change

Constructors

ChangeDoc

Fields

location :: SrcSpan

Location (in the source code) of the changed node.

pprSrc :: SDoc

The PrettyPrint of the changed subnode

pprExec :: SDoc

The PrettyPrint of the new subnode

message :: Maybe String

An optional message to give to the user

category :: ChangeType

A 'type' of Change, which allows ranking them

rewriteSrc :: Outputable node => node -> Change leaf -> Change leaf

Reset the originally changed node.

data ChangeType

Constructors

Wildcard
The Change basically replaces the node with a wildcard. It is not a conclusive change

Removal
A Change that consist in removing a value

Wrapping
A Change that consist of wrapping the actual value E.g. show

Terminal
The Change is good enough to terminate the search and/or be presented to the user as if

(<&&>) :: [Change a] -> (a -> b) -> [Change b]