Seminal.Compiler.Runner

Documentation

runCompiler :: forall a. [FilePath] -> ([(FilePath, ParsedModule)] -> Ghc a) -> IO (Either ErrorMessage a)

Setup and run a GHC Session. The file paths are the paths to the source files to load. The action to run takes the list of loaded modules. Upon error (file access, syntax, ...), throws