Seminal

Description

Entrypoint to Seminal

Documentation

runSeminal :: Options -> [FilePath] -> IO Status

Run Seminal on a source file. If it returns Nothing, the file typechecks, otherwise, provides an ordered list of change suggestions

data Status

Constructors

Success
When the file typechecks without any changes

Error ErrorMessage
When an error occurs during the GHC session, unrelated to typechecking

Changes [(FilePath, ErrorMessage, [Change HsModule])]
For each file that does not typecheck, an ordered list of change suggestions, Along with the original typecheck error