Seminal.Change

Documentation

data Change node

Defines a change to apply on the AST. The namings are inspired by the astRepl (Seminal, 2006, p. 5)

Constructors

Change

Fields

src :: ChangeNode node

exec :: [ChangeNode node]

Run the change, returns the new node

location :: ChangeLocation

followups :: [Change node]

List of subsequent changes to consider, if the change typechecks

message :: String

A User-friendly message to explain why the change worked

category :: ChangeType

Allows ranking the changes.

node :: Outputable n => n -> ChangeNode n

Builds ChangeNode from an AST node

getNode :: ChangeNode n -> n

data ChangeNode nodeType

Sub-data, that could be either the original node, or the change

(<$$>) :: (a -> b) -> [Change a] -> [Change b]

(<&&>) :: [Change a] -> (a -> b) -> [Change b]

show :: ChangeNode node -> ChangeNode node -> ChangeLocation -> String

showWithMessage :: ChangeNode node -> ChangeNode node -> ChangeLocation -> String -> String

data ChangeType

Categories of changes, that allow ordering them

Constructors

Wildcard
The Change basically replaces the node with a wildcard. It is not a conclusive change

Removal
A Change that consist in removing a value

Wrapping
A Change that consist of wrapping the actual value E.g. show

Terminal
The Change is good enough to terminate the search and/or be presented to the user as if

Addition

changeTypes :: [String]

forceRewrite :: Outputable node => Change node -> Change node